topic4/2HourProject-example/smart-travel-planner/requirements.txt
# Smart Travel Planner Dependencies
langchain>=0.1.0
langchain-openai>=0.0.5
langgraph>=0.0.20
requests>=2.31.0
python-dotenv>=1.0.0

# Web Server
flask>=3.0.0
flask-cors>=4.0.0

# Testing
pytest>=7.4.0
pytest-mock>=3.12.0
